feat: migrations

+++ ClubManager/ClubManager.Domain/Entities/Infrastructures/FacilityReservation.cs
@@ -4,18 +4,18 @@ namespace ClubManager.Domain.Entities.Infrastructures
 {
     public class FacilityReservation : BaseEntity
     {
-        public int FacilityId { get; set; }  
+        public long FacilityId { get; set; }  
         public DateTime StartTime { get; set; }  
         public DateTime EndTime { get; set; }
         public string EventType { get; set; }  
         public string EventDescription { get; set; }  
-        public int ReservedUserId { get; set; }  
+        public long ReservedUserId { get; set; }  
         // Relacionamentos com outras tabelas
         public Facility Facility { get; set; }  
         public User ReservedUser { get; set; }
 
 
-        public void SetFacilityId(int facilityId)
+        public void SetFacilityId(long facilityId)
         {
             FacilityId = facilityId;
         }
@@ -40,7 +40,7 @@ namespace ClubManager.Domain.Entities.Infrastructures
             EventDescription = eventDescription;
         }
 
-        public void SetReservedUserId(int reservedUserId)
+        public void SetReservedUserId(long reservedUserId)
         {
             ReservedUserId = reservedUserId;
         }
