+++ ClubManager/ClubManager.Domain/Entities/Infrastructures/MaintenanceHistory.cs
@@ -4,16 +4,16 @@ namespace ClubManager.Domain.Entities.Infrastructures
 {
     public class MaintenanceHistory :BaseEntity
     {
-        public int FacilityId { get; set; }  
+        public long FacilityId { get; set; }  
         public string MaintenanceType { get; set; }  
         public string Description { get; set; }  
         public DateTime MaintenanceDate { get; set; }  
-        public int RequestUserId { get; set; } 
+        public long RequestUserId { get; set; } 
         // Relacionamentos com outras tabelas
         public Facility Facility { get; set; } 
         public User RequestUser { get; set; }
 
-        public void SetFacilityId(int facilityId)
+        public void SetFacilityId(long facilityId)
         {
             FacilityId = facilityId;
         }
@@ -33,7 +33,7 @@ namespace ClubManager.Domain.Entities.Infrastructures
             MaintenanceDate = maintenanceDate;
         }
 
-        public void SetRequestUserId(int requestUserId)
+        public void SetRequestUserId(long requestUserId)
         {
             RequestUserId = requestUserId;
         }
