+++ ClubManager/ClubManager.Domain/Entities/MembersTeams/TeamCoach.cs
@@ -4,18 +4,18 @@ namespace ClubManager.Domain.Entities.MembersTeams
 {
     public class TeamCoach : BaseEntity
     {
-        public int TeamId { get; set; }
+        public long TeamId { get; set; }
         public Team Team { get; set; }
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         public User User { get; set; }
         public bool IsHeadCoach { get; set; }
 
-        public void SetTeamId(int teamId)
+        public void SetTeamId(long teamId)
         {
             TeamId = teamId;
         }
 
-        public void SetUserId(int userId)
+        public void SetUserId(long userId)
         {
             UserId = userId;
         }
