+++ ClubManager/ClubManager.Domain/Entities/MembersTeams/TeamPlayer.cs
@@ -2,17 +2,17 @@
 {
     public class TeamPlayer : BaseEntity
     {
-        public int TeamId { get; set; }
+        public long TeamId { get; set; }
         public Team Team { get; set; }
-        public int PlayerId { get; set; }
+        public long PlayerId { get; set; }
         public Player Player { get; set; }
 
-        public void SetTeamId(int teamId)
+        public void SetTeamId(long teamId)
         {
             TeamId = teamId;
         }
 
-        public void SetPlayerId(int playerId)
+        public void SetPlayerId(long playerId)
         {
             PlayerId = playerId;
         }
