+++ ClubManager/ClubManager.Domain/Entities/MembersTeams/UserClubMember.cs
@@ -4,8 +4,8 @@ namespace ClubManager.Domain.Entities.MembersTeams
 {
     public class UserClubMember : BaseEntity
     {
-        public int UserId { get; set; }       
-        public int ClubMemberId { get; set; }
+        public long UserId { get; set; }       
+        public long ClubMemberId { get; set; }
         public User User { get; set; }          
         public ClubMember ClubMember { get; set; }
     }
